Add type hint to the Factory and move test location
